removed logger from files and changed version of black
Removed logger in files. Outputs are now in stderror or stdout.

Changed the version of black in the .pre-commit-config.yaml back to version 23.7.0 and changed the version of black installed when using pip install -e '.[dev]' to version 23.7.0.

Used black on files and removed imports that are currently not being used in layup command functions.