Predict verb cmd_line args draft (#119)
* Added cmd line arguments for predict verb and fixed bug

Added cmd line arguments for predict verb.
Fixed bug in find_directory_or_exit and its unit tests.

* added some additional command line options

added in some more command line options for layup predict that were in convert and not in the predict verb. Added a check if the user sets the number of days to less than 0.

* Adding a function that will handle convering jd_TDB string or date string into jd_TDB format. Adds logic for start and end date.

* Responding to PR comments. Code organization. Clean up help messages in the cli modules.

* Adding `--station` input, using spiceypy.str2et for time conversion when possible.

* Update CLI flag.

* Removing datetime.strptime parsing. Expanding help message. Added more error handling/reporting.

---------

Co-authored-by: Meg Schwamb <megan.schwamb@gmail.com>
Co-authored-by: Drew Oldag <awoldag@uw.edu>